Refactor Evm.init_module()
