+++ crates/module-system/module-implementations/sov-evm/src/lib.rs
@@ -1,12 +1,11 @@
 #![deny(missing_docs)]
 #![doc = include_str!("../README.md")]
 
+mod account_storage_key;
 mod call;
 mod evm;
 mod genesis;
 mod hooks;
-
-use std::str::FromStr;
 
 pub use call::*;
 pub use evm::*;
@@ -43,11 +42,13 @@ use sov_modules_api::{
     TxState, UnmeteredStateWrapper,
 };
 use sov_state::codec::BcsCodec;
-use sov_state::{EncodeLike, User};
+use sov_state::User;
 
+use crate::account_storage_key::AccountStorageKey;
 use crate::evm::db::EvmDb;
-use crate::evm::primitive_types::SealedBlock;
-use crate::evm::primitive_types::{Block, Receipt, TransactionSignedAndRecovered};
+use crate::evm::primitive_types::{
+    Block, PendingTransaction, Receipt, SealedBlock, TransactionSignedAndRecovered,
+};
 
 // Gas per transaction not creating a contract.
 #[cfg(feature = "native")]
@@ -56,40 +57,6 @@ pub(crate) const MIN_TRANSACTION_GAS: u64 = 21_000u64;
 pub(crate) const MIN_CREATE_GAS: u64 = 53_000u64;
 
 pub use conversions::convert_to_transaction_signed;
-
-/// A pending Ethereum transaction.
-#[derive(Clone, Debug, serde::Serialize, serde::Deserialize)]
-pub struct PendingTransaction {
-    pub(crate) transaction: TransactionSignedAndRecovered,
-    pub(crate) receipt: Receipt,
-}
-
-/// The key to a policy, consisting of the payer and payee addresses with a separator.
-#[derive(Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize, derive_more::Display)]
-#[display(r#"{}/slot/{}"#, self.0, self.1)]
-pub struct AccountStorageKey(pub Address, pub U256);
-
-impl FromStr for AccountStorageKey {
-    type Err = anyhow::Error;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let Some((addr, location)) = s.split_once("/slot/") else {
-            anyhow::bail!("Invalid AccountStorageKey - missing '/slot'' separator");
-        };
-
-        Ok(AccountStorageKey(
-            Address::from_str(addr)?,
-            U256::from_str(location)?,
-        ))
-    }
-}
-
-impl EncodeLike<(&Address, &U256), AccountStorageKey> for BcsCodec {
-    fn encode_like(&self, borrowed: &(&Address, &U256)) -> Vec<u8> {
-        let mut out = self.encode_like(borrowed.0);
-        out.extend_from_slice(&self.encode_like(borrowed.1));
-        out
-    }
-}
 
 /// The sov-evm module provides compatibility with the EVM.
 #[allow(dead_code)]
@@ -315,24 +282,4 @@ impl<S: Spec> Evm<S> {
             .get(tx_hash, state)
             .unwrap_infallible()
     }
-}
-
-#[test]
-fn test_account_storage_key_encode_like() {
-    use sov_state::StateItemEncoder;
-    let key = AccountStorageKey(Address::from_slice(&[1; 20]), U256::from(0));
-    let encoded_like = BcsCodec.encode_like(&(&key.0, &key.1));
-
-    assert_eq!(&BcsCodec.encode(&key), &encoded_like);
-}
-
-#[test]
-fn test_account_storage_key_str_roundtrip() {
-    let key = AccountStorageKey(Address::from_slice(&[1; 20]), U256::from(5));
-    assert_eq!(
-        key.to_string(),
-        "0x0101010101010101010101010101010101010101/slot/5"
-    );
-
-    assert_eq!(AccountStorageKey::from_str(&key.to_string()).unwrap(), key);
 }
